+++ DataIngestion/src/readers/json.jl
@@ -1,7 +1,6 @@
 function json_reader(
-        N::Integer;
+        files::StringList;
         compression::Maybe{AbstractString} = nothing,
-        filename::Bool = false,
         format::AbstractString = "auto",
         hive_partitioning::Bool = false,
         ignore_errors::Bool = false,
@@ -12,7 +11,6 @@ function json_reader(
 
     options = StringDict(
         "compression" => compression,
-        "filename" => filename,
         "format" => format,
         "hive_partitioning" => hive_partitioning,
         "ignore_errors" => ignore_errors,
@@ -21,5 +19,5 @@ function json_reader(
         "union_by_name" => union_by_name,
     )
 
-    return reader_call("read_json", N, options)
+    return reader_call("read_json", files, options)
 end