+++ Pipelines/test/evaluation.jl
@@ -11,7 +11,7 @@
         allows_partition = false,
         allows_weights = false
     )
-    Pipelines.register_wild_card(:trivial; label = "Trivial", settings)
+    Pipelines.register_wild_card(:trivial, "Trivial"; settings)
 
     function trivialcard(inputs::AbstractVector, output::Union{AbstractVector, AbstractString})
         outputs = isa(output, AbstractVector) ? output : [output]
@@ -57,6 +57,8 @@
     (; source_vars, output_vars) = Pipelines.EnrichedDiGraph(nodes)
     @test source_vars == ["temp", "wind"]
     @test output_vars == ["pred humid", "pred wind", "pred temp", "wind name"]
+    @test Pipelines.get_source_vars(Pipelines.Pipeline(nodes)) == source_vars
+    @test Pipelines.get_output_vars(Pipelines.Pipeline(nodes)) == output_vars
 
     faulty_node = Pipelines.Node(trivialcard(["temp"], "pred temp"))
     @test_throws "pred temp" Pipelines.digraph(vcat(nodes, [faulty_node]))
@@ -160,7 +162,7 @@ mktempdir() do dir
             names(df),
             [
                 "No", "year", "month", "day", "hour", "pm2.5", "DEWP", "TEMP",
-                "PRES", "cbwd", "Iws", "Is", "Ir", "_name",
+                "PRES", "cbwd", "Iws", "Is", "Ir",
                 "_percentile_partition", "_tiled_partition",
                 "PRES_rescaled", "TEMP_rescaled",
             ]
@@ -178,7 +180,7 @@ mktempdir() do dir
             names(df),
             [
                 "No", "year", "month", "day", "hour", "pm2.5", "DEWP", "TEMP",
-                "PRES", "cbwd", "Iws", "Is", "Ir", "_name",
+                "PRES", "cbwd", "Iws", "Is", "Ir",
                 "_percentile_partition", "_tiled_partition",
                 "PRES_rescaled", "TEMP_rescaled",
             ]