+++ CHANGELOG.md
@@ -17,6 +17,9 @@
 - `register_wild_card` and `WildCardSettings` are the preferred way to register a wild card [#123](https://github.com/LimenResearch/DashiBoard/pull/123).
 - The API `"method": "m"` + `"method_options": {"opt1": v1, "opt2": v2}` configuration in Pipelines is superseded by `method: {"type": "m", "opt1": v1, "opt2": v2}`. In the StreamlinerCard, the same change occurred for `model` and `training`, and the data `funnel` has now to be passed explicitly in the same way [#143](https://github.com/LimenResearch/DashiBoard/pull/143).
 - In Pipelines, the dot notation for `method_options`, e.g., `"method_options.kmeans.n_classes": 7`, is no longer supported [#143](https://github.com/LimenResearch/DashiBoard/pull/143).
+- When loading several files, the filename is no longer added by default as a new `_name` column, pass `filename = "_name"` to recover the old behavior [#149](https://github.com/LimenResearch/DashiBoard/pull/149).
+- `DuckDBUtils.to_sql` now defaults to `duckdb` dialect [#151](https://github.com/LimenResearch/DashiBoard/pull/151).
+- In `DataIngestion.load_files`, `union_by_name` now respect DuckDB default (i.e., `false`) [#151](https://github.com/LimenResearch/DashiBoard/pull/151).
 
 ### Features
 
@@ -25,3 +28,4 @@
 - Training and evaluation now support callbacks [#99](https://github.com/LimenResearch/DashiBoard/pull/99).
 - Support transformation in `DataIngestion.select` [#104](https://github.com/LimenResearch/DashiBoard/pull/104).
 - `DuckDBUtils.render_params` now can also accepts a `Repository` instead of a `SQLCatalog` [#106](https://github.com/LimenResearch/DashiBoard/pull/106).
+- `DataIngestion.load_files` support loading as a view [#151](https://github.com/LimenResearch/DashiBoard/pull/151).